Fixes description for "The Last Ash Drake" #64424
About The Pull Request

It used to be: "Hunted to extinction, the icey moon of lavaland is the final resting place for these creatures."

I have changed it to: "Hunted to extinction, this icey moon is the final resting place for these creatures."
Why It's Good For The Game

Flows better, and "on lavaland" doesn't exactly make sense when you're on the ice moon. Maybe it's referring to the space ruin the ash drake is in? Still doesn't really sit correctly.
Changelog

cl
spellcheck: The last ash drake currently residing on the celestial body underneath Ice Box Station no longer thinks it's on lavaland.
/cl