Updates the icon of the Legion achievement and adds another for the "Look Out, Sir!" one. (#64624)
The guy that resprited Legion forgot about the achievement, because it's still using the old sprite from the previous decade.

Some of the misc achievements are also still lacking their own icons. I don't have the time right now to work on all of them, but I have sprited one (also as a practice) so to make the task easier. It's a bit smaller than other achievements and more soulless, but that's how I wanted it to be.